Splitted reducer
