+++ src/store/reducers/root-reducer.js
@@ -1,0 +1,14 @@
+import {combineReducers} from "redux";
+
+import {gameData} from "./game-data/game-data";
+import {gameProcess} from "./game-process/game-process";
+
+export const NameSpace = {
+  DATA: `DATA`,
+  GAME: `GAME`,
+};
+
+export default combineReducers({
+  [NameSpace.DATA]: gameData,
+  [NameSpace.GAME]: gameProcess,
+});
